Introduce `to_sqlite_error`
